refactor: get token from url
