+++ packages/access-control-front/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/hooks/use-local-strategy';
 export * from './lib/hooks/use-user-management';
 export * from './lib/hooks/use-metamask-strategy';
+export * from './lib/hooks/use-token-from-url';
